Add script to download and import sample data for #1.
This script allong with the additional `make seed` command downloads and
imports a sample set of data into the postgres database. It replaces all
existing data and uses the default import template for osm2pgsql. Note: It
does not use hstore and so some tag data is lost on import.